+++ src/plone/app/fhirfield/indexes/PluginIndexes/__init__.py
@@ -4,23 +4,7 @@
 # @Link    : http://nazrul.me/
 # @Version : $Id$
 # All imports here
-from .FHIRIndex import FhirActivityDefinitionIndex  # noqa: F401
-from .FHIRIndex import FhirCarePlanIndex  # noqa: F401
-from .FHIRIndex import FhirDeviceIndex  # noqa: F401
-from .FHIRIndex import FhirDeviceRequestIndex  # noqa: F401
 from .FHIRIndex import FhirFieldIndex  # noqa: F401
-from .FHIRIndex import FhirHealthcareServiceIndex  # noqa: F401
-from .FHIRIndex import FhirObservationIndex  # noqa: F401
-from .FHIRIndex import FhirOrganizationIndex  # noqa: F401
-from .FHIRIndex import FhirPatientIndex  # noqa: F401
-from .FHIRIndex import FhirPlanDefinitionIndex  # noqa: F401
-from .FHIRIndex import FhirPractitionerIndex  # noqa: F401
-from .FHIRIndex import FhirProcedureRequestIndex  # noqa: F401
-from .FHIRIndex import FhirQuestionnaireIndex  # noqa: F401
-from .FHIRIndex import FhirQuestionnaireResponseIndex  # noqa: F401
-from .FHIRIndex import FhirRelatedPersonIndex  # noqa: F401
-from .FHIRIndex import FhirTaskIndex  # noqa: F401
-from .FHIRIndex import FhirValueSetIndex  # noqa: F401
 from .FHIRIndex import REGISTRABLE_CLASSES  # noqa: F401